\name{headerSearchPatterns}
\alias{headerSearchPatterns}
\title{headerSearchPatterns
}
\description{
Standard search patterns to analyze result text files produced by the python library particleShearSimulation
}
\usage{
headerSearchPatterns()
}


\value{
Vector of standard regular expression search patterns}

\author{
Thomas Braschler
}


\keyword{ misc }

