\name{read_first_lines}
\alias{read_first_lines}

\title{
read_first_lines
}
\description{
Reads the first lines of a text file
}
\usage{
read_first_lines(path,n=100)
}

\arguments{
  \item{path}{
The path to the file to be read
}
  \item{n}{
The number of lines to be read
}
}

\author{
Thomas Braschler
}

\keyword{ misc }

