\name{split_tensor_and_overview_data}
\alias{split_tensor_and_overview_data}

\title{
split_tensor_and_overview_data
}
\description{
Splits an output file generated by the python simulation into a header part as a small text file and an rda file for the stress tensor data
}
\usage{
split_tensor_and_overview_data(origin_root,path,file_name,destination_root,overwrite=FALSE)
}

\arguments{
  \item{origin_root}{
 Root path for the full output file
}
  \item{path}{
 Path (folders) within the \code{origin_root} location
 }
 \item{file_name}{File name of the text file to be split}

\item{destination_root}{
Destination path. The function expects the folder structure within the destination location to be the same as in the original location; no folders are created by this function.
}
\item{overwrite}{
Should already existing text and rda files be overwritten?
}

}
\details{The original text file will be split into a header text file, having the same file name as the original file but preceded by "header_", and a stress tensor rda file. The file name of the rda file is composed of the prefix "stress_", followed by the origina filename, followed by the extension which is ".rda" instead of ".txt". The aim of the splitting operation is permit more rapid access to the stress tensor data, as reading a text file is substantially slower than reading a .rda file}

\author{
Thomas Braschler
}

\keyword{ misc }

\examples{
\dontrun{
split_tensor_and_overview_data(origin_root=system.file("",package="particleShearEvaluation"),path="","full_python_simulation_output_example.txt")
}
}
